+++ src/db/connection.js
@@ -7,7 +7,7 @@ const mongoose = require("mongoose");
 
 // eslint-disable-next-line no-undef
 mongoose
-    .connect(process.env.MONGODB_URL, {
+    .connect(process.env.MONGO_URL, {
         useNewUrlParser: true,
         useCreateIndex: true,
         useUnifiedTopology: true,